move initialize_runtime to ext/node
