+++ cli/tools/repl/session.rs
@@ -19,6 +19,7 @@ use deno_core::serde_json;
 use deno_core::serde_json::Value;
 use deno_core::LocalInspectorSession;
 use deno_graph::source::Resolver;
+use deno_runtime::deno_node;
 use deno_runtime::worker::MainWorker;
 
 use super::cdp;
@@ -460,8 +461,9 @@ impl ReplSession {
     if !npm_imports.is_empty() || has_node_specifier {
       if !self.has_initialized_node_runtime {
         self.proc_state.prepare_node_std_graph().await?;
-        crate::node::initialize_runtime(
+        deno_node::initialize_runtime(
           &mut self.worker.js_runtime,
+          crate::node::MODULE_ALL_URL.as_str(),
           self.proc_state.options.node_modules_dir(),
         )
         .await?;
